♻️ refactor: 유틸리티 함수 클래스 통합
