+++ src/App.js
@@ -1,10 +1,8 @@
 import { Console } from '@woowacourse/mission-utils';
 import { PRINT_MESSAGE, READ_MESSAGE } from './constants/constants';
 import { validateIsNumber, validateLength } from './utils/validate';
-import { createCarData } from './utils/createCarData';
 import { createRandomNumber } from './utils/createRandomNumber';
-import { goStopCar } from './utils/goStopCar';
-import { pickWinner } from './utils/pickWinner';
+import Utils from './utils/Utils';
 class App {
   async play() {
     const carName = await Console.readLineAsync(READ_MESSAGE.start);
@@ -15,12 +13,12 @@ class App {
 
     Console.print(PRINT_MESSAGE.result);
 
-    let carData = createCarData(carNameList);
+    let carData = Utils.createCarData(carNameList);
 
     for (let i = 0; i < moveCount; i++) {
       carData = createRandomNumber(carData);
 
-      carData = goStopCar(carData);
+      carData = Utils.goStopCar(carData);
 
       carData.forEach((data) => {
         Console.print(`${data.name} : ${data.result}`);
@@ -29,7 +27,7 @@ class App {
       Console.print('\n');
     }
 
-    const winners = pickWinner(carData);
+    const winners = Utils.pickWinner(carData);
 
     Console.print(`${PRINT_MESSAGE.winner}${winners.join(', ')}`);
   }
