--- src/utils/createCarData.js
@@ -1,9 +1,0 @@
-export const createCarData = (carNameList) => {
-  const carData = [];
-
-  carNameList.forEach((carName) => {
-    carData.push({ name: carName, number: 0, result: '' });
-  });
-
-  return carData;
-};
